+++ Cecilifier.Core/Constants.cs
@@ -2,39 +2,34 @@ namespace Cecilifier.Core;
 
 public struct Constants
 {
-    public static ContextFlagsValues ContextFlags = new();
-    public static ParameterAttributesValues ParameterAttributes = new();
-    public static CecilCommonConstants CommonCecilConstants = new();
-    
     public const string AssemblyReferenceCacheBasePath = "/tmp/CecilifierUserAssemblyReferenceCache";
 
-    public struct CecilCommonConstants
+    public struct Cecil
     {
-        public readonly string StaticFieldAttributes = "FieldAttributes.Public | FieldAttributes.Static";
-        public readonly string StaticClassAttributes = "TypeAttributes.AnsiClass | TypeAttributes.Sealed | TypeAttributes.Abstract | TypeAttributes.BeforeFieldInit";
-        public readonly string InterfaceMethodAttributes = "MethodAttributes.SpecialName | MethodAttributes.Public | MethodAttributes.NewSlot | MethodAttributes.Virtual | MethodAttributes.Abstract | MethodAttributes.HideBySig";
-        public readonly string MethodAttributesSpecialName = "MethodAttributes.SpecialName";
-        public readonly string DelegateMethodAttributes = "MethodAttributes.Public | MethodAttributes.HideBySig | MethodAttributes.NewSlot | MethodAttributes.Virtual";
-        public readonly string CtorAttributes = "MethodAttributes.RTSpecialName | MethodAttributes.SpecialName";
+        public const string StaticFieldAttributes = "FieldAttributes.Public | FieldAttributes.Static";
+        public const string StaticTypeAttributes = "TypeAttributes.Abstract | TypeAttributes.Sealed";
+        public const string StaticClassAttributes = $"TypeAttributes.AnsiClass | TypeAttributes.BeforeFieldInit | {StaticTypeAttributes}";
+        public const string InterfaceMethodAttributes = "MethodAttributes.SpecialName | MethodAttributes.Public | MethodAttributes.NewSlot | MethodAttributes.Virtual | MethodAttributes.Abstract | MethodAttributes.HideBySig";
+        public const string MethodAttributesSpecialName = "MethodAttributes.SpecialName";
+        public const string DelegateMethodAttributes = "MethodAttributes.Public | MethodAttributes.HideBySig | MethodAttributes.NewSlot | MethodAttributes.Virtual";
         
-        public readonly string InstanceConstructorName = "ctor";
-        public readonly string StaticConstructorName = "cctor";
+        public const string CtorAttributes = "MethodAttributes.RTSpecialName | MethodAttributes.SpecialName";
+        public const string InstanceConstructorName = "ctor";
+        public const string StaticConstructorName = "cctor";
 
 
-        public CecilCommonConstants() { }
+        public Cecil() { }
     }
     
-    public struct ParameterAttributesValues
+    public struct ParameterAttributes
     {
-        public readonly string None = "ParameterAttributes.None";
-        public readonly string In = "ParameterAttributes.In";
-        public readonly string Out = "ParameterAttributes.Out";
-        public readonly string Optional = "ParameterAttributes.Optional";
-
-        public ParameterAttributesValues() { }
+        public const string None = "ParameterAttributes.None";
+        public const string In = "ParameterAttributes.In";
+        public const string Out = "ParameterAttributes.Out";
+        public const string Optional = "ParameterAttributes.Optional";
     }
     
-    public struct ContextFlagsValues
+    public struct ContextFlags
     {
         /// <summary>
         /// In expressions such as `Index i; i = ^10` the assignment is actually a call to the Index constructor which in IL
@@ -42,13 +37,11 @@ public struct Constants
         /// normal store local, field, parameter). This will take place when we are visiting the lhs of the expression, so
         /// we turn on a flag to indicate that the address of the lhs should be loaded  
         /// </summary>
-        public readonly string PseudoAssignmentToIndex = "pseudo-assignment-to-index";
-        public readonly string HasStackallocArguments = "has-stackalloc-argument";
-        public readonly string RefReturn = "ref-return";
-        public readonly string Fixed = "fixed";
-        public readonly string InRangeExpression = "in-range-expressions";
-
-        public ContextFlagsValues() { }
+        public const string PseudoAssignmentToIndex = "pseudo-assignment-to-index";
+        public const string HasStackallocArguments = "has-stackalloc-argument";
+        public const string RefReturn = "ref-return";
+        public const string Fixed = "fixed";
+        public const string InRangeExpression = "in-range-expressions";
     }
 }